Initial version of app, converted to gulp-based build pipeline!
Incredibly difficult process at this point, but everything seems to be working. I'll
probably come back and clean up this directory structure a bit since I think it'd be
better to do things like install typings as part of `npm install` vs including them
in the repo. Not positive and I'm kinda scared of breaking things again at this point.