Dramatic cleanup; removed old files, eliminated DerivedClock, converted to observables.
This took a lot of work because I had no expose to observables in the past. HOWEVER, the
awesome news is that once I started to wrap my head around them it became clear that I
could reorganize my code to dramatically simplify how I kept track of time. The first
step was to convert the interface between my components and services from promises to
observables, which actually works really well for continuous things like time (its not
too hard to imagine time as a stream of events, which observables requires).

I've been pretty uncomfortable with the division between Clocks and DerivedClocks (as well
as static and dynamic DerivedClocks), which seemed to make things too complicated. Once I
simplified my interface it became clear that I was really just dealing with clocks, some of
which had adjustable offsets. DerivedClocks are no more.

This is more Angular-y, but more importantly I think its actually organized to perform well
for a clock app. Alarms should be pretty straightforward, as should dynamic offsets provided
by periodic polls of the squiggle server.